Use delayed trigger alerts for workout override reminder
